iter15: refactoring
