+++ internal/app/util/util.go
@@ -1,0 +1,18 @@
+package util
+
+import (
+	"math/rand"
+	"time"
+)
+
+func GenerateShortKey() string {
+	const charset = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
+	const keyLength = 6
+
+	r := rand.New(rand.NewSource(time.Now().UnixNano()))
+	shortKey := make([]byte, keyLength)
+	for i := range shortKey {
+		shortKey[i] = charset[r.Intn(len(charset))]
+	}
+	return string(shortKey)
+}
